requirements.txt
apache-airflow-providers-google
pandas
pyarrow
requests
beautifulsoup4
